add signature valid logic to login ,logout, and common operations.

+++ src/main/java/com/syswin/temail/gateway/service/SessionService.java
@@ -1,14 +1,8 @@
 package com.syswin.temail.gateway.service;
 
 
-import java.util.Collection;
-import java.util.function.Consumer;
-
-import javax.annotation.Resource;
-
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.syswin.temail.gateway.TemailGatewayProperties;
-import com.syswin.temail.gateway.encrypt.util.SHA256Coder;
 import com.syswin.temail.gateway.entity.CDTPPacket;
 import com.syswin.temail.gateway.entity.CDTPProtoBuf.CDTPLogin;
 import com.syswin.temail.gateway.entity.CDTPProtoBuf.CDTPLoginResp;
@@ -17,6 +11,9 @@ import com.syswin.temail.gateway.entity.Response;
 import com.syswin.temail.gateway.entity.Session;
 import com.syswin.temail.gateway.exception.PacketException;
 import io.netty.channel.Channel;
+import java.util.Collection;
+import java.util.function.Consumer;
+import javax.annotation.Resource;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -30,17 +27,14 @@ import org.springframework.web.client.RestTemplate;
 public class SessionService {
 
   private final LoginService loginService;
-  private final Consumer<Response<Void>> responseConsumer = ignored -> {
-  };
 
-  private SHA256Coder sha256Coder = new SHA256Coder();
+  private final Consumer<Response<Void>> responseConsumer = ignored -> {};
 
   @Resource
   private ChannelHolder channelHolder;
 
   @Resource
   private RemoteStatusService remoteStatusService;
-
 
   @Autowired
   public SessionService(RestTemplate restTemplate, TemailGatewayProperties properties) {
@@ -59,13 +53,7 @@ public class SessionService {
     } catch (InvalidProtocolBufferException e) {
       throw new PacketException(e, packet);
     }
-
-
-    // TODO(姚华成): 这个cdtpLogin对象暂时没用，后续根据业务需要再完善？
-    // TODO 当前认证请求做简化处理，未来需要完善
-    ResponseEntity<Response> responseEntity = loginService.validSignature(temail,
-        packet.getHeader().getSignature(), extractUnsignedData(packet),
-        packet.getHeader().getSignatureAlgorithm() + "");
+    ResponseEntity<Response> responseEntity = loginService.validSignature(packet);
     Response response = responseEntity.getBody();
     if (responseEntity.getStatusCode().is2xxSuccessful()) {
       loginSuccess(channel, packet, response);
@@ -122,15 +110,20 @@ public class SessionService {
    * @param packet 用户请求数据包
    */
   public void logout(Channel channel, CDTPPacket packet) {
-    // TODO(姚华成) 对packet进行合法性校验
-    String temail = packet.getHeader().getSender();
-    String deviceId = packet.getHeader().getDeviceId();
-    remoteStatusService.removeSession(temail, deviceId, responseConsumer);
-    CDTPLogoutResp.Builder builder = CDTPLogoutResp.newBuilder();
-    builder.setCode(HttpStatus.OK.value());
-    packet.setData(builder.build().toByteArray());
-    channel.writeAndFlush(packet);
-    channelHolder.removeSession(temail, deviceId, channel);
+    ResponseEntity<Response> responseEntity = loginService.validSignature(packet);
+    if (responseEntity.getStatusCode().is2xxSuccessful()) {
+      String temail = packet.getHeader().getSender();
+      String deviceId = packet.getHeader().getDeviceId();
+      remoteStatusService.removeSession(temail, deviceId, responseConsumer);
+      CDTPLogoutResp.Builder builder = CDTPLogoutResp.newBuilder();
+      builder.setCode(HttpStatus.OK.value());
+      packet.setData(builder.build().toByteArray());
+      channel.writeAndFlush(packet);
+      channelHolder.removeSession(temail, deviceId, channel);
+    } else {
+      //TODO - juaihua 如果登出操作验签不通过什么也不做？
+
+    }
   }
 
 
@@ -144,24 +137,5 @@ public class SessionService {
     remoteStatusService.removeSessions(sessions, responseConsumer);
   }
 
-
-  /**
-   * 从CDTPPack中提取签名验证用的原始字符串
-   *    string(commandspace+command)+receiver+string(timestamp)+SHA256(data)
-   */
-  public String extractUnsignedData(CDTPPacket cdtpPacket){
-    StringBuilder unSignedData = new StringBuilder();
-    unSignedData.append((cdtpPacket.getCommandSpace()+cdtpPacket.getCommand()))
-        .append(cdtpPacket.getHeader().getReceiver())
-        .append(cdtpPacket.getHeader().getTimestamp())
-        .append(sha256Coder.encryptAndSwitch2Base64(cdtpPacket.getData()));
-    return unSignedData.toString();
-  }
-
-
-  public static void main(String[] args) {
-    StringBuilder stringBuilder = new StringBuilder();
-    System.out.println(stringBuilder.append(1+1).toString());
-  }
 
 }
